refactor(cli, config): allow selecting test files with any extensions
