+++ models/ConfigFileOptions.ts
@@ -9,10 +9,6 @@ export interface ConfigFileOptions {
      */
     failFast?: boolean;
     /**
-     * The allowed file extension list.
-     */
-    fileExtension?: Array<string>;
-    /**
      * The path to a directory containing files of a test project.
      */
     rootPath?: string;
