--- models/__tests__/__fixtures__/invalid-fileExtension-1.json
@@ -1,4 +1,0 @@
-{
-  "$schema": "../../config-schema.json",
-  "fileExtension": [true]
-}
