--- models/__tests__/__fixtures__/invalid-fileExtension-3.json
@@ -1,4 +1,0 @@
-{
-  "$schema": "../../config-schema.json",
-  "fileExtension": ["json", "json"]
-}
