--- models/__tests__/__fixtures__/valid-fileExtension.json
@@ -1,4 +1,0 @@
-{
-  "$schema": "../../config-schema.json",
-  "fileExtension": ["js", "ts"]
-}
