+++ models/config-schema.json
@@ -6,24 +6,6 @@
       "description": "Stop running tests after the first failed assertion.",
       "type": "boolean"
     },
-    "fileExtension": {
-      "default": [
-        "ts",
-        "tsx",
-        "mts",
-        "cts",
-        "js",
-        "jsx",
-        "mjs",
-        "cjs"
-      ],
-      "description": "The allowed file extension list.",
-      "items": {
-        "type": "string"
-      },
-      "type": "array",
-      "uniqueItems": true
-    },
     "rootPath": {
       "default": "./",
       "description": "The path to a directory containing files of a test project.",
