+++ source/config/ConfigService.ts
@@ -23,7 +23,6 @@ export class ConfigService {
 
   static #defaultOptions: Required<ConfigFileOptions> = {
     failFast: false,
-    fileExtension: ["ts", "tsx", "mts", "cts", "js", "jsx", "mjs", "cjs"],
     rootPath: Path.resolve("./"),
     target: [Environment.typescriptPath == null ? "latest" : "current"],
     testFileMatch: ["**/*.tst.*", "**/__typetests__/*.test.*", "**/typetests/*.test.*"],
@@ -107,11 +106,11 @@ export class ConfigService {
   }
 
   selectTestFiles(): Array<string> {
-    const { fileExtension, pathMatch, rootPath, testFileMatch } = this.resolveConfig();
+    const { pathMatch, rootPath, testFileMatch } = this.resolveConfig();
 
     let testFilePaths = this.compiler.sys.readDirectory(
       rootPath,
-      fileExtension,
+      /* extensions */ undefined,
       /* exclude */ undefined,
       /* include */ testFileMatch,
     );
