--- tests/validation-fileExtension.test.js
@@ -1,55 +1,0 @@
-import { afterEach, describe, test } from "mocha";
-import * as assert from "./__utilities__/assert.js";
-import { clearFixture, getFixtureFileUrl, getTestFileName, writeFixture } from "./__utilities__/fixture.js";
-import { spawnTyche } from "./__utilities__/tstyche.js";
-
-const testFileName = getTestFileName(import.meta.url);
-const fixtureUrl = getFixtureFileUrl(testFileName, { generated: true });
-
-afterEach(async function() {
-  await clearFixture(fixtureUrl);
-});
-
-describe("'fileExtension' configuration file option", function() {
-  test("when option value is not a list", async function() {
-    const config = {
-      fileExtension: "json",
-    };
-
-    await writeFixture(fixtureUrl, {
-      ["tstyche.config.json"]: JSON.stringify(config, null, 2),
-    });
-
-    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl);
-
-    assert.equal(stdout, "");
-
-    await assert.matchSnapshot(stderr, {
-      fileName: `${testFileName}-wrong-option-value-type-stderr`,
-      testFileUrl: import.meta.url,
-    });
-
-    assert.equal(exitCode, 1);
-  });
-
-  test("when item of the list is not a string", async function() {
-    const config = {
-      fileExtension: ["json", true],
-    };
-
-    await writeFixture(fixtureUrl, {
-      ["tstyche.config.json"]: JSON.stringify(config, null, 2),
-    });
-
-    const { exitCode, stderr, stdout } = await spawnTyche(fixtureUrl);
-
-    assert.equal(stdout, "");
-
-    await assert.matchSnapshot(stderr, {
-      fileName: `${testFileName}-wrong-list-item-type-stderr`,
-      testFileUrl: import.meta.url,
-    });
-
-    assert.equal(exitCode, 1);
-  });
-});
